added a global dependencies object for infra module
