--- src/config/index.ts
@@ -1,7 +1,0 @@
-import RoutesMap from "./route_map";
-import { ROUTES } from "./routes";
-
-export {
-    RoutesMap,
-    ROUTES
-}
