+++ src/infra/external_dependencies/index.ts
@@ -1,0 +1,6 @@
+/*
+**  Not sure if this is the right pattern to initialize dependencies that infra module have on the client
+*/
+const ExternalDependencies: any = {};
+
+export default ExternalDependencies;
