--- src/infra/screen_renderer/vault_screen/render_item.ts
@@ -1,7 +1,0 @@
-const renderItem = ({ item }: { item: LayoutElement, index: number }) => {
-    const itemComponentProps = dataStore[item.id];
-    const ItemComponent = PropsComponentMap[item.type]
-    return (
-        <ItemComponent {...itemComponentProps} />
-    )
-}
